Updated create Troop form'
