+++ package.json
@@ -19,9 +19,8 @@
     "@react-native-async-storage/async-storage": "~1.15.0",
     "@react-native-community/datetimepicker": "4.0.0",
     "@react-native-community/hooks": "^2.8.1",
-    "@react-native-community/masked-view": "0.1.10",
-    "@react-native-community/picker": "1.6.0",
     "@react-native-community/slider": "4.1.12",
+    "@react-native-picker/picker": "2.2.1",
     "@react-navigation/bottom-tabs": "^5.0.6",
     "@react-navigation/drawer": "^5.5.0",
     "@react-navigation/native": "^5.5.1",
